third_party/WebKit/LayoutTests/fast/dom/shadow/distributed-node-focus-expected.txt


PASS document.activeElement.id is "input1"
PASS document.activeElement.id is "input1"
PASS shadowRoot.activeElement is null
PASS document.activeElement.id is "distributed-node"
PASS shadowRoot.activeElement.id is "distributed-node"
PASS document.activeElement.id is "distributed-node"
PASS shadowRoot.activeElement is null
PASS successfullyParsed is true

TEST COMPLETE
